Don't rely on CloseClientConnections which is unpredictable
The reason it is unpredictable is the lock monitor only retries the request if
it gets back a 500. Closing a connection isn't considered a 500 error,
therefore will cause the lock to be lost. But, the reason this has worked so
far is that http.Get might retry the request under the hood given that some
requirements are met. This unpredictable behavior is believed to what caused
the failure.

The fix is to make the http handler block on the channel and to tighten the
check to use the url query parameters to properly differentiate between the
lock monitor requests and other requests

[#132625243]