Fix lock expiration
Previous change that modified lock pick would cancel an inflight
goroutine if the ModifiedIndex of the infight check was less than or
equal to the one being registered.

This made it so that when the burglar was registering all the locks, it
would continually cancel inflight expiration checks for locks that had
not been incremented. Integration tests were passing because they used a
TTL of 1 second, which expired before the burglar had time to do its
first RegisterTTL loop.

[#142583253]